+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#1](https://github.com/zfcampus/zf-asset-manager/pull/1) fixes how the
+- [zfcampus/zf-asset-manager#1](https://github.com/zfcampus/zf-asset-manager/pull/1) fixes how the
   `public/.gitignore` file is populated, ensuring no duplicates are created.
 
 ## 1.0.0 - 2016-07-26